use jdbc spring to connect to the database and retrieve it

+++ src/main/java/com/springjdbc/dao/ContactDao.java
@@ -8,7 +8,7 @@ public interface ContactDao {
 	List<Contact> findAll();
 	List<Contact> findByFirstName(String firstName);
 	String findLastNameById(Long id);
-	String findFirstNameById(Long id);
+	String findFirstNameById(int id);
 	void insert(Contact contact);
 	void update(Contact contact);
 	void delete(Long contactId);
